v0.0.7版本: 1. 修复 bug; 2. 添加 web 项目最小配置;

+++ agtms-parent/agtms-web/src/main/java/net/saisimon/agtms/web/dto/resp/NavigationInfo.java
@@ -16,7 +16,7 @@ public class NavigationInfo extends BaseInfo {
 	
 	private String icon;
 	
-	private Integer priority;
+	private Long priority;
 	
 	private Long parentId;
